Unify CType & Type
This change makes all CTypes representable by a Type. This allows us to
track type information for global symbols correctly.

There are some small changes in the test outputs, they fall into the
following categories:

- Pointers to primitive typedefs are no longer preserved
    - Ex: `(int *)` is now uniformly `(s32 *)`
- There are now more `(u8)` and `(u16)` explicit casts.
- The format of the `--dump-typemap` was changed to match how variables
are declared. This was a side effect of fixing how the types of function
pointers are printed
- For now, I've allowed locals & temps to be a CType. This means that a
local/temp could be larger than a word, which leads to local vars
"overlapping" on the stack. I've added detection for this for now, which
inserts a warning comment. I imagine a future change replace these
aliased local vars with the correct struct access.
    - This does sometimes lead to emitting incorrect C statements, like
    `sp40 = (Vec3s) sp60;` where only the first word would be copied
    instead of the whole struct.